Login changes (#1069)
* use 2-step onboarding, some refactoring

* show login page any time there is no user_login field in config (meaning user is not authorized)

* Update github copy

---------
