+++ client/src/locales/en.json
@@ -200,7 +200,7 @@
 	"Something went wrong": "Something went wrong",
 	"chats in bloop": "chats in bloop",
 	"Setup bloop": "Setup bloop",
-	"Please log into your GitHub account to complete setup": "Please log into your GitHub account to complete setup",
+	"Please log into your GitHub account to complete setup, this helps us us combat misuse.": "Please log into your GitHub account to complete setup, this helps us us combat misuse.",
 	"Select color theme:": "Select color theme:",
 	"Connect account": "Connect account",
 	"Continue": "Continue",
@@ -416,5 +416,6 @@
 	"Search pages": "Search pages",
 	"This page is currently unavailable. Ability to generate will be resumed as soon as this issue is resolved.": "This page is currently unavailable. Ability to generate will be resumed as soon as this issue is resolved.",
 	"Hold <1></1> to add multiple files": "Hold <1></1> to add multiple files",
-	"Clear section": "Clear section"
+	"Clear section": "Clear section",
+	"Let’s get you started with bloop!": "Let’s get you started with bloop!"
 }
