+++ client/src/pages/Onboarding/index.tsx
@@ -1,31 +1,15 @@
-import React, {
-  memo,
-  useCallback,
-  useContext,
-  useEffect,
-  useState,
-} from 'react';
-import { useTranslation } from 'react-i18next';
-import { useLocation } from 'react-router-dom';
+import React, { memo, useCallback, useContext, useEffect } from 'react';
 import { UIContext } from '../../context/uiContext';
 import { DeviceContext } from '../../context/deviceContext';
-import NavBar from '../../components/NavBar';
 import {
-  getJsonFromStorage,
   getPlainFromStorage,
   ONBOARDING_DONE_KEY,
   REFRESH_TOKEN_KEY,
-  saveJsonToStorage,
   savePlainToStorage,
-  SESSION_ID_KEY,
-  USER_DATA_FORM,
 } from '../../services/storage';
-import { getConfig, getRepos, saveUserData } from '../../services/api';
-import SeparateOnboardingStep from '../../components/SeparateOnboardingStep';
-import StatusBar from '../../components/StatusBar';
-import UserForm from './UserForm';
-import FeaturesStep from './FeaturesStep';
-import SelfServeStep from './SelfServeStep';
+import { getConfig, refreshToken } from '../../services/api';
+import SelfServe from './SelfServe';
+import Desktop from './Desktop';
 
 export type Form = {
   firstName: string;
@@ -35,41 +19,14 @@ export type Form = {
 };
 
 const Onboarding = ({ activeTab }: { activeTab: string }) => {
-  const { t } = useTranslation();
-  const [form, setForm] = useState<Form>({
-    firstName: '',
-    lastName: '',
-    email: '',
-    emailError: null,
-    ...getJsonFromStorage(USER_DATA_FORM),
-  });
-  const [shouldShowPopup, setShouldShowPopup] = useState(false);
   const { shouldShowWelcome, setShouldShowWelcome } = useContext(
     UIContext.Onboarding,
   );
-  const { isGithubConnected, refreshToken } = useContext(
-    UIContext.GitHubConnected,
-  );
-  const { isSelfServe, os, setEnvConfig, envConfig } =
-    useContext(DeviceContext);
-  const location = useLocation();
+  const { isSelfServe, setEnvConfig, envConfig } = useContext(DeviceContext);
 
   const closeOnboarding = useCallback(() => {
     setShouldShowWelcome(false);
     savePlainToStorage(ONBOARDING_DONE_KEY, 'true');
-  }, []);
-
-  useEffect(() => {
-    if (import.meta.env.ONBOARDING) {
-      if (
-        getPlainFromStorage(SESSION_ID_KEY) !==
-        window.__APP_SESSION__.toString()
-      ) {
-        localStorage.removeItem(ONBOARDING_DONE_KEY);
-        savePlainToStorage(SESSION_ID_KEY, window.__APP_SESSION__.toString());
-        setShouldShowWelcome(true);
-      }
-    }
   }, []);
 
   useEffect(() => {
@@ -96,88 +53,29 @@ const Onboarding = ({ activeTab }: { activeTab: string }) => {
         setShouldShowWelcome(true);
       }
     } else {
-      getConfig()
-        .then((d) => {
-          setEnvConfig(d);
-          if (!d.user_login) {
-            setShouldShowWelcome(true);
-          } else {
-            closeOnboarding();
-          }
-        })
-        .catch(() => {
-          setShouldShowWelcome(true);
-        });
+      getConfig().then((d) => {
+        setEnvConfig(d);
+      });
     }
   }, []);
 
-  const onSubmit = useCallback(() => {
-    saveUserData({
-      email: form.email,
-      first_name: form.firstName,
-      last_name: form.lastName,
-      unique_id: envConfig.tracking_id || '',
-    });
-    saveJsonToStorage(USER_DATA_FORM, form);
-    closeOnboarding();
-    setTimeout(() => setShouldShowPopup(true), 1000);
-  }, [form, envConfig.tracking_id]);
+  useEffect(() => {
+    if (!isSelfServe) {
+      if (!envConfig.user_login) {
+        setShouldShowWelcome(true);
+      } else {
+        closeOnboarding();
+      }
+    }
+  }, [envConfig]);
 
   return shouldShowWelcome ? (
     isSelfServe ? (
-      <div className="text-label-title">
-        <NavBar isSkeleton activeTab={activeTab} />
-        <div
-          className={`flex justify-center items-start mt-8 w-screen overflow-auto relative h-[calc(100vh-4rem)]`}
-        >
-          <div className="fixed top-0 bottom-0 left-0 right-0 mt-8 mb-16 bg-bg-base z-80">
-            <div className="absolute top-0 bottom-0 left-0 right-0 flex justify-center items-center overflow-auto">
-              <div className="flex flex-col items-center max-w-md2 w-full">
-                <div className="bg-bg-base border border-bg-border rounded-lg shadow-high p-6 flex flex-col flex-1 gap-8 w-full max-w-md2 relative max-h-[calc(100vh-12rem)]">
-                  <SelfServeStep />
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
-        <StatusBar />
-      </div>
+      <SelfServe activeTab={activeTab} />
     ) : (
-      <div className="fixed top-0 bottom-0 left-0 right-0 z-100 bg-bg-sub select-none">
-        {os.type === 'Darwin' && <NavBar isSkeleton activeTab={activeTab} />}
-        <img
-          src="/light.png"
-          alt=""
-          className="fixed -top-68 lg:-top-80 xl:-top-96 w-[90vw] lg:w-[80vw] xl:w-[69vw] right-0 pointer-events-none opacity-[0.16] z-50"
-        />
-        <div className="flex h-full justify-center mt-8">
-          <div className="w-full lg:w-1/2 h-full flex justify-center">
-            <div
-              className={`w-[512px] h-full flex flex-col items-center justify-center px-13 gap-6`}
-            >
-              <UserForm form={form} setForm={setForm} onContinue={onSubmit} />
-            </div>
-          </div>
-          <div
-            className={`w-1/2 h-full hidden lg:flex justify-center items-center border-l border-bg-border relative 
-        before:absolute before:top-0 before:bottom-0 before:left-0 before:right-0 before:bg-[url('/grainy-pattern.png')] 
-        before:bg-repeat before:mix-blend-soft-light before:opacity-[0.14]`}
-          >
-            <div className="w-[585px]">
-              <img className="onboarding-chats-img" alt={t('chats in bloop')} />
-            </div>
-          </div>
-        </div>
-      </div>
+      <Desktop activeTab={activeTab} closeOnboarding={closeOnboarding} />
     )
-  ) : (
-    <SeparateOnboardingStep
-      isVisible={shouldShowPopup}
-      onClose={() => setShouldShowPopup(false)}
-    >
-      <FeaturesStep handleNext={() => setShouldShowPopup(false)} />
-    </SeparateOnboardingStep>
-  );
+  ) : null;
 };
 
 export default memo(Onboarding);
